pyproject.toml
[tool.poetry]
name = "grobidmetaanalytics"
version = "0.1.0"
description = "Generación de reporte de características de Publicaciones con Grobid y Python"
authors = ["Rubén Cid <rubencid001@gmail.com>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fpdf2 = "^2.7.8"
wordcloud = "^1.9.3"
matplotlib = "^3.8.3"
numpy = "^1.26.4"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
